requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
langchain==0.1.0
langchain-google-genai==0.0.6
langgraph==0.0.20
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
google-generativeai==0.3.1
aiofiles==23.2.1
openai
langchain